requirements.txt
aiofiles==23.1.0
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
asttokens==2.2.1
bcrypt==4.0.1
bidict==0.22.1
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
click-log==0.4.0
colorama==0.4.6
cryptography==41.0.2
deprecation==2.1.0
dill==0.3.6
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-socketio==0.0.10
gitdb==4.0.10
GitPython==3.1.32
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
invoke==1.7.3
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
keyring==24.2.0
lazy-object-proxy==1.9.0
markdown2==2.4.9
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==9.1.0
mypy==1.4.1
mypy-extensions==1.0.0
netifaces==0.11.0
nicegui==1.3.0
numpy==1.25.1
onetimepad==1.4
orjson==3.9.2
packaging==23.1
paramiko==3.2.0
parso==0.8.3
pkginfo==1.9.6
platformdirs==3.9.1
postgrest==0.10.6
pscript==0.7.7
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.3.0
Pygments==2.15.1
pylint==2.17.4
PyNaCl==1.5.0
pyserial==3.5
pysqlitecipher==0.22
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-gitlab==3.15.0
python-multipart==0.0.6
python-semantic-release==7.33.2
python-socketio==5.8.0
pywin32-ctypes==0.2.2
PyYAML==6.0
readme-renderer==40.0
realtime==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
storage3==0.5.2
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
thonny==4.1.1
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
vbuild==0.8.1
watchfiles==0.19.0
webencodings==0.5.1
websockets==10.4
wrapt==1.15.0
zipp==3.16.2
